DbImporter bug fix: execute last query even if < 500; reduced appointment and person test data

+++ src/main/java/ch/bfh/bti7081/s2016/purple/HealthVisitor/data/businesslogic/DbImport.java
@@ -56,6 +56,9 @@ public abstract class DbImport {
     		String line1 = "";
     		
     		while ((line = br.readLine()) != null) {
+    			if (line == "")
+    				continue;
+    			
     			if (count == 0){
     				if (deleteFirst){
     					query += "DELETE FROM " + tableName + "; ";
@@ -83,6 +86,11 @@ public abstract class DbImport {
     		      	count = 0;
     			}
 			}
+    		
+    		if (query.length() > 0){
+    			stmt = conn.createStatement();
+		      	stmt.executeUpdate(query);
+    		}
     		
 	      	logger.debug("import from " + fileName + " into " + tableName + " successful");
     	} catch(SQLException se){
